+++ chartkick.js
@@ -2,7 +2,7 @@
  * Chartkick.js
  * Create beautiful JavaScript charts with minimal code
  * https://github.com/ankane/chartkick.js
- * v1.4.2
+ * v1.5.0
  * MIT License
  */