+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added Chart.js adapter **beta**
 - Added `smarterDates` option (temporary until 2.0)
+- Added `smarterDiscrete` option (temporary until 2.0)
 - Fixed line height on timeline charts
 
 ## 1.4.2